pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "butterfence"
version = "0.3.2"
description = "Claude Code-native safety harness that red-teams repos against destructive/secret-leaking agent behavior"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pyyaml>=6",
    "pathspec>=0.11",
    "anthropic>=0.39",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.5"]
api = ["fastapi>=0.109", "uvicorn>=0.27", "pydantic>=2"]
pdf = ["reportlab>=4"]
edge = ["numpy>=1.24", "onnx>=1.14", "onnxruntime>=1.16"]
dev = ["pytest>=8", "httpx>=0.27"]

[project.scripts]
butterfence = "butterfence.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/butterfence"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
